Ignore empty child nodes in the DOM on new browser versions
Starting from Chomium 117, a new "Pretty Print" toggle is included adding a new empty div in the DOM to do this action. Because of that, the pre node was instead an empty div node. By filtering on the content length, we will find the pre, or the raw html if no headers were set.